chore: merge all of the speller's vocabs into a single file
